refactor: add submodules and restructuration of the code
