+++ rest-api/src/main/java/com/linkauto/restapi/RestApiApplication.java
@@ -1,0 +1,14 @@
+package com.linkauto.restapi;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class RestApiApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(RestApiApplication.class, args);
+	}
+
+}
+
