+++ rest-api/src/main/java/com/linkauto/restapi/repository/UserRepository.java
@@ -1,0 +1,14 @@
+package com.linkauto.restapi.repository;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.linkauto.restapi.model.User;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, String> {
+    Optional<User> findByUsername(String username);
+    boolean existsByUsername(String username);
+}
