+++ rest-api/src/main/resources/dbsetup.sql
