+++ rest-api/src/main/resources/static/styles.css
