+++ rest-api/src/test/java/com/linkauto/restapi/RestApiApplicationTests.java
@@ -1,0 +1,13 @@
+package com.linkauto.restapi;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class RestApiApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
