--- src/main/java/com/example/restapi/RestApiApplication.java
@@ -1,14 +1,0 @@
-package com.example.restapi;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class RestApiApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(RestApiApplication.class, args);
-	}
-
-}
-
