--- src/main/java/com/example/restapi/repository/PostRepository.java
@@ -1,10 +1,0 @@
-package com.example.restapi.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.example.restapi.model.Post;
-
-@Repository
-public interface PostRepository extends JpaRepository<Post, Long> { 
-}
