--- src/main/java/com/example/restapi/repository/UserRepository.java
@@ -1,14 +1,0 @@
-package com.example.restapi.repository;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.example.restapi.model.User;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, String> {
-    Optional<User> findByUsername(String username);
-    boolean existsByUsername(String username);
-}
