--- src/main/resources/application.properties
@@ -1,22 +1,0 @@
-#spring.application.name=rest-api
-#spring.datasource.url=jdbc:h2:mem:testdb
-#spring.datasource.driverClassName=org.h2.Driver
-#spring.datasource.username=root
-#spring.datasource.password=password
-#spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
-#spring.h2.console.enabled=true
-
-# Spring Datasource Settings for MySQL
-spring.datasource.url=jdbc:mysql://localhost:3306/linkautodb?allowPublicKeyRetrieval=true&useSSL=false
-spring.datasource.username=la
-spring.datasource.password=la
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-
-# JPA Settings
-spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=true
-
-springdoc.api-docs.path=/api-docs
-springdoc.api-docs.enabled=true
-springdoc.swagger-ui.path=/swagger-ui.html
