--- src/test/java/com/example/restapi/RestApiApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.restapi;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class RestApiApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
